ActionRecognitionModule/python/requirements.txt
numpy
pandas
scikit-learn
pyserial
